requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
Pillow==10.4.0
